+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* TODO: Creating package to show floating widget and allow see in full screen the same widget.